py calculator/requirements.txt
zeroc-ice>=3.7.0
kafka-python>=2.0.2
pydantic>=2.0.0
python-dotenv>=1.0.0
confluent-kafka